fix: rename platform.py to vllm_platform.py to avoid stdlib shadowing
